Decouple validation and error parsing
